requirements.txt
alabaster==0.7.12
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
autopep8==1.6.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
biopython==1.79
bleach==5.0.1
certifi==2021.10.8
cffi==1.15.1
chardet==4.0.0
click==8.1.2
configparser==5.2.0
cycler==0.10.0
debugpy==1.6.3
decorator==4.4.2
defusedxml==0.7.1
-f https://data.dgl.ai/wheels/repo.html
dgl-cu111==0.8.1
dglgo==0.0.1
docker-pycreds==0.4.0
docutils==0.17.1
edlib==1.3.9
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.1
gitdb==4.0.9
GitPython==3.1.27
idna==2.10
imagesize==1.3.0
importlib-metadata==4.11.3
importlib-resources==5.9.0
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.0.1
json5==0.9.10
jsonschema==4.14.0
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyter_client==7.3.5
jupyterlab==3.4.5
jupyterlab-pygments==0.2.2
jupyterlab_server==2.15.1
kiwisolver==1.3.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.3
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.5.1
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.20.3
numpydoc==1.2.1
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
Pygments==2.12.0
pyparsing==2.4.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.1
requests==2.25.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==0.24.2
scipy==1.6.3
Send2Trash==1.8.0
sentry-sdk==1.5.10
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==4.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.5.0
subprocess32==3.5.4
terminado==0.15.0
threadpoolctl==2.1.0
tinycss2==1.1.1
toml==0.10.2
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cu111
-f https://download.pytorch.org/whl/torch_stable.html
torchaudio==0.9.0
-f https://download.pytorch.org/whl/torch_stable.html
torchvision==0.10.0+cu111
tornado==6.2
tqdm==4.62.2
traitlets==5.3.0
typer==0.4.1
typing-extensions==3.10.0.0
urllib3==1.26.5
wandb==0.12.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.0
zipp==3.8.0
